Refs #152 - Don't use proxy for SendEmailCommand.
