+++ src/Synapse/Email/SendEmailCommand.php
@@ -3,13 +3,15 @@
 namespace Synapse\Email;
 
 use Synapse\Command\CommandInterface;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 use LogicException;
 use OutOfBoundsException;
 
-class SendEmailCommand implements CommandInterface
+class SendEmailCommand extends Command
 {
     /**
      * @var EmailMapper
@@ -85,5 +87,18 @@ class SendEmailCommand implements CommandInterface
         }
 
         $output->writeln('Email sent successfully!');
+    }
+
+    /**
+     * Set name, description, arguments, and options for this console command
+     */
+    protected function configure()
+    {
+        $this->setDescription('Send an email')
+            ->addArgument(
+                'id',
+                InputArgument::REQUIRED,
+                'ID of email to send'
+            );
     }
 }
